update contact-form plugins

+++ contact-form.php
@@ -7,15 +7,19 @@ Version: 1.0
 Author: Alfian SS
 Author URI: http://w3guy.com
 */
-global $con_db_version;
-$con_db_version = '1.0';
-add_option( "con_db_version", "1.0" );
+
+
 
 class Contact_Form {
+
+	public $con_db_version = '1.0';
+
+	public function con_db_version() {
+		return $this->con_db_version;		
+	}
 		
-	public static function con_install() {
+	public function con_install() {
 		global $wpdb;
-		global $con_db_version;
 	
 		$table_name = $wpdb->prefix . 'testimonial';
 		
@@ -34,56 +38,32 @@ class Contact_Form {
 		dbDelta( $sql );
 	
 		// add_option( 'con_db_version', $con_db_version );
-		update_option( 'con_db_version', $con_db_version );
-	}
-	
-	public static function con_install_data() {
-		global $wpdb;
-		
-		$welcome_name = 'Mr. WordPress';
-		$welcome_text = 'Congratulations, you just completed the installation!';
-		
-		$table_name = $wpdb->prefix . 'testimonial';
-		
-		$wpdb->insert( 
-			$table_name, 
-			array( 
-				'time' => current_time( 'mysql' ), 
-				'name' => $welcome_name, 
-				'text' => $welcome_text, 
-			) 
-		);
+		update_option( 'con_db_version', $this->con_db_version() );
 	}
 
-	public static function form_contact_update_db_check() {
-		global $con_db_version;
-		if ( get_site_option( 'con_db_version' ) != $con_db_version ) {
-			self::con_install();
+	public function form_contact_update_db_check() {
+		if ( get_option( 'con_db_version' ) != $this->con_db_version() ) {
+			$this->con_install();
 		}
 	}
 
-	public static function delete_testimonial() {
+	public function delete_testimonial() {
 		global $wpdb;
 		
-		if(isset($_GET['delete'])) {
+		if(isset($_GET['delete']) && isset($_GET['id'])) {
 			$id = $_GET['id'];
 
 			$table = $wpdb->prefix.'testimonial';			
 			$result = $wpdb->delete( $table, array( 'id' => $id ) );
 
-			if($result) {
-				// wp_redirect('');
-				// exit;
-			} else {
-
-			}
+			
 		}
 
 	}
 
-	public static function contact_form_menu(){
+	public function contact_form_menu(){
 		echo "<div class='wrap'><h2>Manage Contact Form</h2></div>";
-		// $this->delete_testimonial();
+		$this->delete_testimonial();
 		global $wpdb;
 		$results = $wpdb->get_results( "SELECT * FROM {$wpdb->prefix}testimonial", OBJECT );
 
@@ -104,7 +84,7 @@ class Contact_Form {
 				<td>".$tm->phone_number."</td>
 				<td>".$tm->testimoni."</td>
 				<td>
-					<a href='admin.php?page=contact-form&delete=true&id=".$tm->id."' onclick='".self::delete_testimonial()."'>Delete</a>
+					<a href='".admin_url('admin.php?page=contact-form&delete=true&id='.$tm->id)."'>Delete</a>
 				</td></tr>";
 			$no++;
 		}
@@ -112,11 +92,11 @@ class Contact_Form {
 		echo "</table>";
 	}
 
-	public static function cf_admin_menu() {
-		add_menu_page( 'Contact Form Menu', 'Contact Form', 'manage_options', 'contact-form', array('Contact_Form','contact_form_menu'), 'dashicons-tickets', 6  );
+	public function cf_admin_menu() {
+		add_menu_page( 'Contact Form Menu', 'Contact Form', 'manage_options', 'contact-form', array($this,'contact_form_menu'), 'dashicons-tickets', 6  );
 	}
 
-	public static function save_testimonial() {
+	public function save_testimonial() {
 		// if the submit button is clicked, send the email
 		if ( isset( $_POST['cf-submitted'] ) ) {
 			global $wpdb;
@@ -147,23 +127,23 @@ class Contact_Form {
 		
 	}
 
-	public static function html_form_code() {
+	public function html_form_code() {
 		include( 'templates/forms.php' );
-		self::save_testimonial();
+		$this->save_testimonial();
 	}
 
-	public static function cf_shortcode() {
+	public function cf_shortcode() {
 		ob_start();    
-		self::html_form_code();    
+		$this->html_form_code();    
 		return ob_get_clean();
 	}
 }
 
 $contact_form = new Contact_Form();
+// $contact_form::con_db_version();
 require_once( plugin_dir_path( __FILE__ ) . 'class.contact-form-widget.php' );
 
 register_activation_hook( __FILE__, array($contact_form,'con_install') );
-register_activation_hook( __FILE__, array($contact_form,'con_install_data') );
 
 add_action( 'plugins_loaded', array($contact_form,'form_contact_update_db_check') );
 add_action( 'admin_menu', array($contact_form,'cf_admin_menu') );
